Remove _Interface
