+++ Spatial_Engine/Modify/ISetInternalElements2D.cs
