+++ Spatial_Engine/Modify/ISetOutlineElements1D.cs
