+++ Spatial_Engine/Query/IElements0D.cs
