+++ Spatial_Engine/Query/IInternalElements2D.cs
