+++ Spatial_Engine/Query/IOutlineElements1D.cs
